Implement task duration measurement and logging enhancement
A new TimeUtility class is added to calculate the duration of tasks in seconds. It is used in the calculateTax method of the TaxCalculationService to measure the time taken by the tax calculation operation. This enhancement in logging will help in performance monitoring and future optimization. Logs also include the duration of each operation upon completion.